chore: move migration folder to src
